Rename zluda_dump to zluda_trace
