+++ Cargo.toml
@@ -23,13 +23,13 @@ members = [
     "zluda_blaslt",
     "zluda_common",
     "zluda_dnn",
-    "zluda_dump",
-    "zluda_dump_blas",
-    "zluda_dump_blaslt",
-    "zluda_dump_common",
-    "zluda_dump_dnn",
-    "zluda_dump_fft",
-    "zluda_dump_sparse",
+    "zluda_trace",
+    "zluda_trace_blas",
+    "zluda_trace_blaslt",
+    "zluda_trace_common",
+    "zluda_trace_dnn",
+    "zluda_trace_fft",
+    "zluda_trace_sparse",
     "zluda_fft",
     "zluda_inject",
     "zluda_ml",
