--- zluda_dump_fft/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "zluda_dump_fft"
-version = "0.0.0"
-authors = ["Andrzej Janik <vosen@vosen.pl>"]
-edition = "2021"
-
-[lib]
-name = "zluda_dump_fft"
-crate-type = ["cdylib"]
-
-[dependencies]
-format = { path = "../format" }
-dark_api = { path = "../dark_api" }
-cuda_macros = { path = "../cuda_macros" }
-cuda_types = { path = "../cuda_types" }
-zluda_dump_common = { path = "../zluda_dump_common" }
-libloading = "0.8"
-paste = "1.0"
-unwrap_or = "1.0.1"
-cglue = "0.3.5"
-
-[package.metadata.zluda]
-linux_symlinks = [
-    "dump/libcufft.so",
-    "dump/libcufft.so.11"
-]
