+++ zluda_ml/Cargo.toml
@@ -16,6 +16,6 @@ cuda_types = { path = "../cuda_types" }
 linux_symlinks = [
     "libnvidia-ml.so",
     "libnvidia-ml.so.1",
-    "dump/libnvidia-ml.so",
-    "dump/libnvidia-ml.so.1",
+    "trace/libnvidia-ml.so",
+    "trace/libnvidia-ml.so.1",
 ]
