+++ zluda_trace/src/dark_api.rs
