+++ zluda_trace_common/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "zluda_trace_common"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+
+[dependencies]
+libloading = "0.8"
+cuda_types = { path = "../cuda_types" }
+dark_api = { path = "../dark_api" }
+format = { path = "../format" }
+cglue = "0.3.5"
